added object oriented programming
added structs for components
